+++ posthog-react-native/src/hooks/usePostHog.ts
@@ -1,6 +1,6 @@
 import { PostHog } from '../posthog-rn'
 import React from 'react'
-import { PostHogContext } from '../PostHogContext'
+import { PostHogContext } from '../PostHogContext__'
 
 export const usePostHog = (): PostHog => {
   const { client } = React.useContext(PostHogContext)
